Refactoring: filter for disabled scanners from tool_type table
