+++ dojo/filters.py
@@ -2,6 +2,8 @@ __author__ = 'Jay Paz'
 import collections
 import logging
 from datetime import timedelta, datetime
+from functools import reduce
+
 from django import forms
 from django.apps import apps
 from auditlog.models import LogEntry
@@ -32,7 +34,7 @@ from dojo.test.queries import get_authorized_tests
 from dojo.finding.queries import get_authorized_findings
 from dojo.endpoint.queries import get_authorized_endpoints
 from django.forms import HiddenInput
-
+from dojo.tools.factory import get_disabled_scanners
 logger = logging.getLogger(__name__)
 
 local_tz = pytz.timezone(get_system_setting('time_zone'))
@@ -46,6 +48,14 @@ def custom_filter(queryset, name, value):
     filter = ('%s__in' % (name))
     return queryset.filter(Q(**{filter: values}))
 
+
+def manage_disabled_scanners():
+    disabled = get_disabled_scanners()
+    # this dirty/ugly as hell solution is to filter out scanners that are disabled, needs some refactory on design level
+    q_list = []
+    q_list = map(lambda n: Q(name__iexact=n), disabled)
+    q_list = reduce(lambda a, b: a | b, q_list)
+    return q_list
 
 def now():
     return local_tz.localize(datetime.today())
